% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSTP.R
\name{BSTP}
\alias{BSTP}
\title{Fit a Bayesian Spatiotemporal Poisson model}
\usage{
BSTP(y,X,A,nt,
            nchain=3,nsim=100,nburn=20,nthin=1)
}
\arguments{
\item{y}{vector of counts, must be non-negative}

\item{X}{matrix of covariates, numeric}

\item{A}{adjacency matrix, numeric}

\item{nt}{positive integer, number of time points}

\item{nchain}{positive integer, number of MCMC chains to be run}

\item{nsim}{positive integer, number of iterations in each chain}

\item{nburn}{non-negative integer, number of iterations to be discarded as burn-in samples}

\item{nthin}{positive integer, thinning interval}
}
\value{
list of posterior samples of the parameters of the model
}
\description{
Generate posterior samples for the parameters in a Bayesian Spatiotemporal Poisson Model
}
