% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BZINB.R
\name{BZINB}
\alias{BZINB}
\title{Fit a Bayesian Zero Inflated Negative Binomial Model}
\usage{
BZINB(y,Xtilde,A, oind=NULL,
             nchain=3,niter=1000,nburn=500,nthin=1)
}
\arguments{
\item{y}{vector of counts, must be non-negative}

\item{Xtilde}{matrix of offset and covariates, numeric}

\item{A}{adjacency matrix, numeric}

\item{oind}{indices of offset}

\item{nchain}{positive integer, number of MCMC chains to be run}

\item{niter}{positive integer, number of iterations in each chain}

\item{nburn}{non-negative integer, number of iterations to be discarded as burn-in samples}

\item{nthin}{positive integer, thinning interval}
}
\value{
list of posterior samples of the parameters of the model
}
\description{
Generate posterior samples for the parameters in a Bayesian Zero Inflated Negative Binomial Model
}
