% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/county.adjacency.R
\docType{data}
\name{county.adjacency}
\alias{county.adjacency}
\title{county.adjacency: A dataframe containing neighborhood information for counties in the US}
\format{
\subsection{\code{county.adjacency}}{

A dataframe with 22200 rows and 4 columns
}
}
\usage{
county.adjacency
}
\description{
Data set containing neighborhood information for counties in the US, to be used to create adjacency matrices
}
\keyword{datasets}
