% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_func.R
\name{qRankPar}
\alias{qRankPar}
\title{Bar plot for time-averaged log-q estimates over quantile-representative counties (descending order)}
\usage{
qRankPar(state.set,ns,nt,cname,stfit,vn=12)
}
\arguments{
\item{state.set}{character vector of set of states on which the the graphics is to be made}

\item{ns}{positive integer, the number of counties}

\item{nt}{positive integer, the number of timepoints}

\item{cname}{character vector of the names of the counties}

\item{stfit}{the fitted data for BSTP, BSTNB or BSTZINB}

\item{vn}{positive integer, number of sample counties to display}
}
\value{
bar graph
}
\description{
Produce a descending order of bar plot for time-averaged log-q estimates over quantile-representative counties
}
