% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_func.R
\name{ResultTableSummary}
\alias{ResultTableSummary}
\title{Summary Table for a fitted object}
\usage{
ResultTableSummary(bstfit)
}
\arguments{
\item{bstfit}{fitted object using the function BSTP, BSTNB or BSTZINB}
}
\value{
summary table
}
\description{
Generates a short summary table for a fitted object using BSTP, BSTNB or BSTZINB function
}
