% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdat.R
\docType{data}
\name{simdat}
\alias{simdat}
\title{simdat: A simulated dataset containing response and covariates with region and time information}
\format{
\subsection{\code{simdat}}{

A dataframe with 2376 rows and 5 columns: sid (region ID), tid (timepoint), y (count response), V1 (intercept), and x (covariate).
}
}
\usage{
simdat
}
\description{
Synthetic dataframe to be used for examples and trial runs
}
\keyword{datasets}
