% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth_dat_IA.R
\docType{data}
\name{synth_dat_IA}
\alias{synth_dat_IA}
\title{synth_dat_IA: A dataset containing county-wise synthsized counts for the state of Iowa over 16 time points and a covariate used to generate the data}
\format{
\subsection{\code{synth_dat_IA}}{

A dataframe with 99 rows and 4 columns: sid (ID number for each county in Iowa where the response was recorded), tid (Time point when the response were recorded), y (Response) and x (Covariate used to generate the data).
}
}
\usage{
synth_dat_IA
}
\description{
Dataframe to be used to check the functions work properly on synthetic data
}
\keyword{datasets}
