% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/USAcities.R
\docType{data}
\name{USAcities}
\alias{USAcities}
\title{USAcities: A dataset containing state and county information for the cities in the United States}
\format{
\subsection{\code{USAcities}}{

A dataframe with 3232 rows and 4 columns: state_id (State abbreviation), county_name (County name), county_fips (FIPS codes for the counties) and population (County population).
}
}
\usage{
USAcities
}
\description{
Dataframe to be used internally to make maps and get county information
}
\keyword{datasets}
