% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSTZINB.R
\name{BSTZINB}
\alias{BSTZINB}
\title{Fit a Bayesian Spatiotemporal Zero Inflated Negative Binomial model}
\usage{
BSTZINB(y,X,A,LinearT = TRUE,
            nchain=3,niter=100,nburn=20,nthin=1)
}
\arguments{
\item{y}{vector of counts, must be non-negative}

\item{X}{matrix of covariates, numeric}

\item{A}{adjacency matrix, numeric}

\item{LinearT}{logical, whether to fit a linear or non-linear temporal trend}

\item{nchain}{positive integer, number of MCMC chains to be run}

\item{niter}{positive integer, number of iterations in each chain}

\item{nburn}{non-negative integer, number of iterations to be discarded as burn-in samples}

\item{nthin}{positive integer, thinning interval}
}
\value{
list of posterior samples of the parameters of the model
}
\description{
Generate posterior samples for the parameters in a Bayesian Spatiotemporal Zero Inflated Negative Binomial Model
}
