% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_func.R
\name{compute_ZINB_DIC}
\alias{compute_ZINB_DIC}
\title{DIC for BSTZINB fitted objects}
\usage{
compute_ZINB_DIC(y,fit,lastit,nchain)
}
\arguments{
\item{y}{vector of counts, must be non-negative, the response used for fitting a BSTZINB model}

\item{fit}{BSTZINB fitted object}

\item{lastit}{positive integer, size of the chain used to fit BSTZINB}

\item{nchain}{positive integer, number of chains used to fit BSTZINB}
}
\value{
DIC value
}
\description{
Computes DIC for a BSTZINB fitted object
}
