% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_func.R
\name{ResultTableSummary2}
\alias{ResultTableSummary2}
\title{Generate a summary table of the outputs all different methods given the data}
\usage{
ResultTableSummary2(y,X,A,nt,
                           nchain=3,nsim=100,nburn=20,nthin=1)
}
\arguments{
\item{y}{vector of 0s and 1s or two categories (will be coerced to 0-1)}

\item{X}{matrix of covariates, numeric}

\item{A}{adjacency matrix, numeric}

\item{nt}{positive integer, number of time points}

\item{nchain}{positive integer, number of MCMC chains to be run}

\item{nsim}{positive integer, number of iterations in each chain}

\item{nburn}{non-negative integer, number of iterations to be discarded as burn-in samples}

\item{nthin}{positive integer, thinning interval}
}
\value{
summary tables for the different methods
}
\description{
Fits BSTP, BSTNB and BSTZINB (with linear or non-linear temporal trend) to a given data and summarizes the results in a table
}
